+++ mayan/settings/development/ddt.py
@@ -1,4 +1,4 @@
-from .development import *  # NOQA
+from .base import *  # NOQA
 
 # Stop debug toolbar patching!
 # see https://github.com/django-debug-toolbar/django-debug-toolbar/issues/524