+++ mayan/settings/development/docker/db_postgres.py
@@ -1,0 +1,12 @@
+from .base import *  # NOQA
+
+DATABASES = {
+    'default': {
+        'ENGINE': 'django.db.backends.postgresql_psycopg2',
+        'NAME': 'postgres',
+        'USER': 'postgres',
+        'HOST': 'localhost',
+        'PASSWORD': 'mayanuserpass',
+        'PORT': '5432',
+    }
+}